+++ frontend/src/pages/recommend/components/NewsletterHero/NewsletterHero.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { useQuery } from '@tanstack/react-query';
 import { Link } from '@tanstack/react-router';
-import SlideCardList from '../../SlideCardList/SlideCardList';
+import SlideCardList from '../SlideCardList/SlideCardList';
 import { queries } from '@/apis/queries';
 import logo from '#/assets/logo.png';
