+++ frontend/src/pages/recommend/components/TrendySection/TrendySection.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useQuery } from '@tanstack/react-query';
 import { useState } from 'react';
 import { createPortal } from 'react-dom';
-import NewsletterDetail from '../../NewsletterDetail/NewsletterDetail';
+import NewsletterDetail from '../NewsletterDetail/NewsletterDetail';
 import { queries } from '@/apis/queries';
 import Chip from '@/components/Chip/Chip';
 import ImageInfoCard from '@/components/ImageInfoCard/ImageInfoCard';
